# perf: memorize exclude function in webpack config