@ijjk 2025-07-11
## Failing test suites <!--41536289729-->
Commit: 9449c3e18440af81a653226f9b10da4089da830c

`pnpm test test/integration/invalid-page-automatic-static-optimization/test/index.test.js`(rspack)
- Invalid Page automatic static optimization > production mode > handles non-error correctly

<details>
<summary>Expand output</summary>

  ● Invalid Page automatic static optimization › production mode › handles non-error correctly

    expect(received).toMatch(expected)

    Expected pattern: /invalid API token/
    Received string:  " ⚠ Linting is disabled.
    Failed to compile.·
    ../../../packages/next/dist/build/webpack/loaders/next-route-loader/index.js?kind=PAGES&page=%2Falso-invalid&preferredRegion=&absolutePagePath=private-next-pages%2Falso-invalid.js&absoluteAppPath=next%2Fdist%2Fpages%2F_app&absoluteDocumentPath=next%2Fdist%2Fpages%2F_document&middlewareConfigBase64=e30%3D!
      × Cannot find module 'private-next-pages/also-invalid.js' for matched aliased key 'private-next-pages'·
    ../../../packages/next/dist/build/webpack/loaders/next-route-loader/index.js?kind=PAGES&page=%2Falso-valid&preferredRegion=&absolutePagePath=private-next-pages%2Falso-valid.js&absoluteAppPath=next%2Fdist%2Fpages%2F_app&absoluteDocumentPath=next%2Fdist%2Fpages%2F_document&middlewareConfigBase64=e30%3D!
      × Cannot find module 'private-next-pages/also-valid.js' for matched aliased key 'private-next-pages'·
    ../../../packages/next/dist/build/webpack/loaders/next-route-loader/index.js?kind=PAGES&page=%2Finvalid&preferredRegion=&absolutePagePath=private-next-pages%2Finvalid.js&absoluteAppPath=next%2Fdist%2Fpages%2F_app&absoluteDocumentPath=next%2Fdist%2Fpages%2F_document&middlewareConfigBase64=e30%3D!
      × Cannot find module 'private-next-pages/invalid.js' for matched aliased key 'private-next-pages'·
    ../../../packages/next/dist/build/webpack/loaders/next-route-loader/index.js?kind=PAGES&page=%2Fvalid&preferredRegion=&absolutePagePath=private-next-pages%2Fvalid.js&absoluteAppPath=next%2Fdist%2Fpages%2F_app&absoluteDocumentPath=next%2Fdist%2Fpages%2F_document&middlewareConfigBase64=e30%3D!
      × Cannot find module 'private-next-pages/valid.js' for matched aliased key 'private-next-pages'··
    > webpack config.resolve.alias was incorrectly overridden. https://nextjs.org/docs/messages/invalid-resolve-alias
    "

      52 |         try {
      53 |           const { stderr } = await nextBuild(appDir, [], { stderr: true })
    > 54 |           expect(stderr).toMatch(/invalid API token/)
         |                          ^
      55 |           expect(stderr).not.toMatch(/without a React Component/)
      56 |         } finally {
      57 |           await fs.remove(path.join(appDir, 'pages'))

      at Object.toMatch (integration/invalid-page-automatic-static-optimization/test/index.test.js:54:26)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).