@bgw 2025-07-11
Woah, this is wild. I've got a few more ideas if you're willing to try them:

- Join all the `babelIncludeRegexes` into a single regex by concatenating them with `|` operators. This is probably the most expensive check here.
- Move the `node_modules` check to the top. That's the cheapest check and is probably the most likely to be false. I don't think this can be merged with `babelIncludeRegexes`?
- Convert `finalTranspilePackages` into a `Set` (maybe not needed here as it looks like by default it has only one element) and make it the last check (since there's only one thing in there by default, so it's least likely to short-circuit the conditional).

If we do those basic optimizations, I wonder if caching is still worth it or not.