@SyMind 2025-07-11
@bgw Based on my investigation, the cache is necessary.

The key bottleneck is the `isResourceInPackages` function. When I tested joining all `babelIncludeRegexes`, no significant improvement was observed. Note that JavaScript's `Array.some()` method already implements short-circuit evaluation by design.

<img width="634" height="204" alt="Performance metrics" src="https://github.com/user-attachments/assets/e9a5b8a1-4226-49f9-8191-d47569150c10" />

Moving the `node_modules` check to the top of the logic chain theoretically improves performance, which is why I've made this change.